Changes following PR:
- Rename `allUntilFailure` to the more positive `allWhileOk`
- Broke out some of the logic from `compose.js` into a new `schema.js` (contains previous code from `internal.js`)
- Added a decidated test file for `schema.js`
