+++ test/validators.spec.js
@@ -9,6 +9,7 @@ import {
   validateIsArray,
   validateArrayItemsHaveType,
   validateArrayItemsPass,
+  alwaysErr,
 } from '../src/lib/validators';
 
 describe('validators', () => {
@@ -220,6 +221,24 @@ describe('validators', () => {
     it('should return OK if all the items pass the given validator', () => {
       const res = validate([4, 5, 6]);
       expect(isOK(res)).to.be.true;
+    });
+  });
+
+  describe('#alwaysErr()', () => {
+    it('should return a validator that always errors', () => {
+      const validate = alwaysErr(['Error one', 'Error two']);
+
+      expect(isErr(validate())).to.be.true;
+      expect(isErr(validate({ test: '1234' }))).to.be.true;
+      expect(isErr(validate([1, 2, 3]))).to.be.true;
+    });
+
+    it('should return a the error passed into it', () => {
+      const errors = ['Error one', 'Error two'];
+
+      const validate = alwaysErr(errors);
+
+      expect(validate().errors).to.deep.equal(errors);
     });
   });
 });
